Create Odd Sum .cpp
Given an integer N, consider all arrays A of size N such that:

All the elements are non-negative and distinct.
All prefix sums are odd. Formally, for all i such that 1≤i≤N, ∑ij=1Ai is odd.
Among all possible arrays A, output the smallest possible sum of the elements of the array.

Note: Since the Input/Output may be large, it is preferred to use fast I/O.

Input Format
The first line contains T - the number of test cases. Then the test cases follow.
The first line of each test case contains N - the size of the array.
Output Format
For each test case, output on one line the smallest sum among all arrays satisfying the constraints.

Constraints
1≤T≤106
1≤N≤109
Sample Input 1 
1
3
Sample Output 1 
3
Explanation
Test case 1: A possible array is [1,2,0]. [1,0,0] is not valid because 0 occurs twice in it; [0,1,2] is not valid because the prefix sum until the first index is 0, which is even. Another possible array is [5,2,4]. [1,2,0] yields the sum 3, and we can prove that there are no valid arrays that have sum less than 3.